Fix SuperMicro badfish calls.
This addresses a lack of exclusion of SuperMicro systems when running
the badfish library to lookup what would be a non-existent
idrac_interfaces.yml entry for "foreman" for one-time boot.  This is not
needed on SuperMicro systems.

This also removes the try/accept here that would normally try to run the
badfish call against non-Dells (which will always fail) and instead just
doesn't try.

We're also changing the list of SUPERMICROS to SUPERMICRO in config.py
so we are sure to reference things like is_supermicro(host):

Lastly, we are going to remove ReturnFalse from this because while fatal
in the grand scheme of provisioning (one system didn't PXE/kickstart) we
don't want to hold up the others - our systems validation should still
tell us that one or more systems are marked for build but this way we
aren't wasting valuable provisioning time.

fixes: https://github.com/redhat-performance/quads/issues/238
Change-Id: Ifd394f645419482c2a4247cf525888bc72e54dc9