Fix SuperMicro badfish calls.
This addresses a lack of exclusion of SuperMicro systems when running
the badfish library to lookup what would be a non-existent
idrac_interfaces.yml entry for "foreman" for one-time boot.  This is not
needed on SuperMicro systems.

This also removes the try/accept here that would normally try to run the
badfish call against non-Dells (which will always fail) and instead just
doesn't try.

We're also changing the list of SUPERMICROS to SUPERMICRO in config.py
so we are sure to reference things like is_supermicro(host):

fixes: https://github.com/redhat-performance/quads/issues/238
Change-Id: Ifd394f645419482c2a4247cf525888bc72e54dc9