minor fix for file perms for simple_table_web
This issue is similar to the one for instackenv generation.
File permissions in python3 should be listed in octal using
the notation : 0o644, not 644.  Otherwise the file permissions
will appear as:

-=>>ls -la /var/www/html/visual/
total 11112
drwxr-xr-x. 2 root root     201 May 10 15:21 .
drwxr-xr-x. 4 root root      33 Mar 11 20:45 ..
--w----r-T. 1 root root  669217 Mar 31 23:50 2019-03.html
--w----r-T. 1 root root 1816779 Apr 30 23:50 2019-04.html
--w----r-T. 1 root root 1860189 May 10 15:20 2019-05.html
[...]

This is minor and may go unnoticed depending on how the
web server is configured.

Change-Id: Ia4d66d8d640dedec37ff4017d1abbfaa91adb22d