broken and unmanaged hosts show up in both
when a host is marked as broken but also not defined in quads,
it will show up in the both lists.  This is because the print_unmanaged
function takes a list of the management interface hosts.  But our
foreman does not define the mgmt-<fqdn> hostname as an independent
foreman entity.  We need to check the broken state of the real_host
and not the manage interface hostname.

Change-Id: I128df388ca61dfc817d4567f678f699d2cd27af9