Fix for move and rebuild execution plus failure notification.
Turns out move and rebuild was failing on Badfish for expecting
different patterns on job search.
This failure led to the provisioned flag skip the rest of the hosts.
Failure notification were not sent neither due to the cloud object
in mongo to never be set as provisioned=True which was a filter
for the validation.

Fixes: https://github.com/redhat-performance/quads/issues/255
Change-Id: I745002546d84e21c5decbf00dd6814b64ea3b2a3