Fix for badfish uncaught exception.
The way it should work is that if it fails on M&R for one host then
it stops execution and moves to the next host on the same cloud.
Just realized that there is an uncaught exception from Badfish that
might stop the whole execution of M&R which would cause it to be
stuck on a cloud if it is failing on a host.

Change-Id: Ia4e42bbc1ad28ad20c926678e88bad4049ba1c37