Fix for default cloud on ls-schedules.
The implementation for default cloud was set to always be cloud01.
Now default-cloud on define-host is mandatory and populates a new
field on the Host document.
Added logic to M&R for hosts to go back to default cloud after an
assignment is over.
Replaced a few calls to REST API for mongo resources.

Fixes:
https: //github.com/redhat-performance/quads/issues/247
https: //github.com/redhat-performance/quads/issues/249
Change-Id: I05b8c651c9d933d368f398a7d13daa98f9d99d8a