document mod_passenger tuning.
Since we're now using asyncio we had to adjust our MaxKeepAliveRequests
by x2 in order to have a healthy amount of simultaneous connections for
large amounts of systems.

Because the semaphore settings were also lowered this seems like an
adequate setting for now.

This documents the above and also scoots the Foreman-specific setup
topics for and under the Foreman section of docs/switch-host-setup.md

Change-Id: I93677ce70266726b63862e4446e5996d3d8920fc