Added tool for fixing foreman permissions.
It might be possible that M&R is not succesfully updating all of
foreman parameters and some times cloud users might be missing
roles.
This tool will bring all current schedules and reassign the cloud
roles based on this.
This will be better used once we have granularity checks on the host
completion and we can detect this not being populated.

Change-Id: I6acd542894f858fd0dd1bae0b1f514f99d641bfb