allow setting first vlan in config
sw_vlan_first can be set to the first VLAN ID used in the first
interface of cloud01.  If not set it defaults to vlan id 1100.

This is useful when you have multiple quads environments that use
interconnected supporting switch infrastructure and you want to
ensure that different environments don't use the same base VLAN
for the various cloud environments.  For example, if you have
"stage" vs "production" and you want to start the VLAN ranges for
the two environments using different ranges.

Change-Id: I542083d30ca18f6857ce10d02839a98ca9f3aaaa