fix: wiki not displaying cloud history
when generating the assignments page we were using the data from
the schedule in that day. Turns out the schedule  object maps to
the cloud object which is constantly being modified as we do for
defining new clouds.
added a fix to look for the first cloudhistory object with the same
cloud name, that has been created right before the schedule object.

closes: https://github.com/redhat-performance/quads/issues/297
Change-Id: I2431a868d539a4636407ae96505119b66cab4f36