feat/fix: make qinq easier, fix behavior
This changes the argparse behavior of --qinq so that the default if not
specified is to use qinq:0 mode or "separated" interfaces. Internally
in mongo this is a "True" value for the qinq attribute.

When --qinq is specified during cloud definition, it will instead set
things up to use qinq:1 mode or "combined" interfaces VLAN design.

Additionally the --ls-qinq output is made to be clearer.

Change-Id: Idbe53a80073fb5fe6fc4544464e645dc13e4774f