Error in calculating the next four months out
there was a logic error in how we look ahead from the current date.
Instead of looking ahead the number of days that exist in the CURRENT
month, we need to look ahead the number of days in NEXT month. This
generally is only an issue when calculating the allocations at the
end of months, in which the following month has fewer days then the
current month.  For example, if we run the code to calculcate the
heat maps on the 31st of March, we would end up skipping April, because
31 days after March 31st becomes April 1, and that would be foolish.

Fixes: https://github.com/redhat-performance/quads/issues/308

Change-Id: I702a376ec8562001ffb2d2761aca641270bd9259