fix: cloud validation with all hosts ready validated
in the scenario of all hosts in a cloud been validated but the
cloud not being updated, it would fall into an endless loop of
validation failures due to self.hosts being empty.
if all hosts have already been validated we can ommit running these
host bounded validations for validating the cloud.

Change-Id: I994d879887a774c2a043fa8d629f71c81e22ae1a