fix: VSC keeps trying to setup public VLAN
for setting public vlans we were retrying on last nic, added a check
for if old_vlan==new_vlan.
fixed minor issue on ls-interface displaying ip instead of mac.
fixed pexpect sripts to avoid waiting for exit since child.close is
already dealing with that.
added warning to validate_env for when grace period not expired.

Change-Id: Iad030637cfe9dbffe9c293bf8b931554055ab1fd