Add Cloud Number and Ticket to Notifications.
This should add cloud numbers and cloud ticket to release and expiration
notifications to make things easier to track and manage.

This also updates updates the "message" template to make it more clear
how to request an extension.

Lastly, we will not append http to our message templates, instead the
full url variables can be done in /opt/quads/conf/quads.yml

fixes: https://github.com/redhat-performance/quads/issues/321
Change-Id: I578e377712088768b8120047d953c75f7bd4892c