feat: readability improvment to notifications
We've already improved the notification subject elements to be more
easily distinguishable and searchable in email.  We should do the same
for the pre_initial messages we send out when an environment is defined
but it's not yet scheduled to go out.

We'll revamp all notification methods here so they identify at least the
cloud environment and ticket number assopciated with them.

Change-Id: Ie979489e7bb02da2c9c3349c43644995f8551c04