Add an inventory file for OpenSHift Baremetal Deployments
OpenShift Baremetal deployments require two MACs to be specified
per host. The current instackenv.json provides only one MAC (provisioning)
per host. It would be possible to add another mac into the list of macs
for each host in instackenv.json itself, but it is unclear how this affects
existing OpenStack deployment automation (especially with SuperMicros).
This commit adds the ability to serve another inventory file, with just a
minor modification of having two MACs per host. The adavntage with having a
separate ocpinventory.json file is that automation could consume it and save time
as well as we now have the ability to further enhance this file in the future
to accomodate more OpenShift specific items. This commit acts as a bridge between
QUADS and automation to deploy OpenShift on Baremetal.

Change-Id: Ib88ee205d59dab627d18aa952f4f809f312e744d