make ocinventory display all macaddr.
Make the generated ocinventory.json generate all the internal
macaddreses from the interface list.

This way folks can consume this and parse whatever interface index(s)
they need and it can be more useful.

This should hopefully replace the need for people ssh'ing to Dell idrac
as well to obtain this information and speed things up.

Change-Id: I7d270e7d1786bd89d8f9c4fd5742c9515637d19e