Allow a conf for not resetting wipe
In some cases where it might be needed to leave the wipe value
alone for vacated cloud allocations, administrators can use the
conf setting:

wipe_reset: False

This may be useful when wanting to test schedules only, and/or
making rapid changes.  One such scenario is to define multiple
cloud environments with wipe=false, and hosts are set to move
back and forth between environments without regard for resetting
hosts (useful when testing switch config changes).  Without this,
the wipe setting of the clouds will revert to true which may not
be desired in the given test scenario.  Use with caution.

Change-Id: Ie0209411bed8874f269534f0fe110c0e783956c3