feat: hosts metadata
Host metadata can be updated via yaml file with:
`bin/quads-cli --define-host-details --metadata conf/hosts_metadata.yml`

Example host metadata yaml included under conf/hosts_metadata.yml

There is no id for the disks, hence updates are not possible, existing records
are removed, what is in the yml is replicated in the mongodb.
If host metadata is not defined we add those values without force, if already
existing, records are deleted and recreated.
Host model is free-form but a list with allowed values is now defined on config.
Accepted operators are: ==, !=, <, <=, >, >=

eg.:

--filter "disks.disk_type==nvme,disks.count>2,disks.size_gb>2000,model==R620"

This patch does not include logic for an `OR` type of search.

Change-Id: I8a4ed4e077ff5fe71c66dd6814e6972f1f99a6a0