Feat: add quads_request_deadline_day
This adds an configuration option for quads_request_deadline_day which
if populated will template expiration notices with the preferred day at
which tenants should submit their extensions.

Also updating reference to "RT" in visual tooltip to be more generic
"Ticket" as it's a better fit and ticket-backend agnostic.

Lastly, small change to our cron examples so validate_env.py is logged
somewhere.  We should have logger library do this for us however
ultimately to include move-and-rebuild logs and just set this as a
configurable path location in /opt/quads/conf/quads.yml

Change-Id: I341ec11ff75ec4108f03bc2f51dde1000e2b817a