fix: Bandaid for validate_env on for build host to reboot
the error message we get ("Could not run reboot for: $host") might
be misleading. the try/except that returns this message is also
attempting to do a boot-to-type via badfish and that is the action
that would fail since there are no jobs to be created for a reboot
action.

Change-Id: I81915b45c8dff389b107443bd5c88c265b396332