feat: added netcat health checks
with netcat we can check for port 22 to be open as a health check
instead of just waiting for redfish to report that the host is up.
this allows quads to wait for the lifecycle controller to complete
whatever task is executing.
added to M&R process plus validate_env for reporting the list of
hosts which are not responding on port 22.

Change-Id: Iee15b06fd48fc269cc241b326071661eeb1082a7