feat: added virtual media unmount
porting over the changes necessary for virtual media check and
virtual media unmount.
Having this to be fail safe as not failing the M&R process if unable
to unmount so it is not disruptive.
This is in place to execute only for supermicros but since it is done
via badfish, we now require to create a badfish instance for SMs too,
which we didn't in the past as it was not required since we use
ipmitool for setting the PXE flag. This will result in an additional
execution time cost for M&R when running against SMs.

Change-Id: I6293583ab9b850a978f3a7e2e880f1243c5509a2