feat: Added shrink and moved extend-cloud to only be `extend`
This will require us to pass either `cloud` or `host`
Also added `--now` argument to be used in place of weeks under the
shrink command for early terminations.
Included with this also `--date` for a specific termination or
extension date.
Grouped all the time bound arguments as `--weeks`, `--months`,
`--year` and `--date` in a mutually exclusive fashion as well as, in
a separate group, `--host` and `--cloud`.

Change-Id: Ife7ec6db6a9411967d99873a506bd750266faf9e