fix: set power state on DL360
when sending a reset instruction to redfish, if the server is
already in that reset type state, the error code thrown by
the DL360s differs from all other and instead of 409 it returns
a 400 error.
This was causing BF to throw an exception and cut the execution
of M&R short.

Change-Id: I55008105eb5b93aa1a3bda1c1175d3dd2834ded3