fix: eerroneous error on VSC nochange mode
When you run VSC against an environment that has a public
vlan associated with it, the vlan_member is not something
that can be calculated on the last nic because the switch
configs on a juniper switch will not have a "set vlans"
matching section.  Therefore the vlan_member variable is
simply set to 0.  The error message should not be thrown
in this case when we are working with the last_nic.

Fixes: https://github.com/redhat-performance/quads/issues/372
Change-Id: I124c9ef770ce351e6a0266ceba3c3910bb3d7d0b