Make --cloud be optional and alert on mismatch
1.  make the use if --cloud not be mandatory.
    However if --cloud is omitted on place of --host,
    we should be able to calculate the cloud object
    based on the hostname provided.

2.  if BOTH --host and --cloud are specified, we should
    do some checking to make sure the target host is actually
    a member of the target cloud.

Under rare circumstances the user may want to reconfigure the
switch ports associated with host and place the host into the
vlans associated with an arbitrary cloud. However this is a rare
corner case. In the case where the user specifies both cloud and
host, and the cloud does not match the associated cloud of the
target host, the user should be alerted with output on the screen
to alert them that there is a possible unintentional mismatch.

Fixes: https://github.com/redhat-performance/quads/issues/378
Change-Id: I9aa804ebcdeaec97b31d7b0de000d8101c7d225d