feat: Tool/example for mapping VLAN ID/mapping.
New tool ls_switch_conf.py allows for a quick listing of VLAN id
mappings per cloud, optionally printing the full configuration and
mapping.

We are also providing a psuedo KOL for quickly figuring out the
VLAN ids based on the cloud for people not on a future packaged version
of QUADS (likely 1.1.6) which would contain this tool and otherise
aren't pulling it directly from the git source repo.

Co-authored-by: grafuls@redhat.com

Change-Id: Ieb0008adaa13dcf4b9bc4989f2be84faed76e50e