fix: Don't trigger VLAN int check if value is null.
This is a fix for the mongoDB value return of 'null' which is
set via quads-cli --mod-cloud --vlan none (or similar).

Otherwise, we trigger the sanity check which tries to convert the value
of _args.vlan from a string into an integer.

Feel free to commandeer this patchset for a better fix, but this works
in STAGE.

fixes:  https://github.com/redhat-performance/quads/issues/392

Change-Id: I6d84a8a132ebbbcfe70ba60c138c0768294888d0