Update cron suggestions to add timestamp
The suggested cron entries for quads are updated in this
patch to place a line along with a date stamp for better
understanding of when the log entries are correlated.  Note
that it is still left to the person in charge of the deployment
to change /dev/null to a log file under /var/log if they want
to capture the output.

Change-Id: I0f8aa00b34b6616a78dfd43a9895702a19f0ced8