feat: notify submitter if watchers are missing
closes: https://github.com/redhat-performance/quads/issues/407
With this accompanying change to Landing https://gitlab.cee.redhat.com/perfscale-infra/landing/-/merge_requests/33
should users whose emails fail to mapped to Jira users on request for access be notfied otf that.
On extension and expansion requests, if we fail to add watchers from the parent ticket, we notify
the submitter of that request.

Change-Id: I57c33be716c7f0ee376ed56e0d1332b26a0b8b55