Do a better check for hosts in build
There is a corner case in which if the foreman templates have a line such as:

sshpw --username=root my_kickstart_password

and the host is marked for build (and may be in the middle of provisioning) quads validate_env.py mistakes the kickstart as complete because validate_env.py only uses Netcat() to determine that the host is reachable. A more comprehensive check is needed to know if the host is reachable using a well known password, and a better determination can be made to know that the host may be mid-kickstart.

Fixes: https://github.com/redhat-performance/quads/issues/445
Change-Id: Ib8e7c36d34c3d224f1fce405fe10101e5bf2de34