Fix systemd Restart= pattern matching.
Since we have multiple gunicorn WSGI processes now we can no longer use
restart=pkill gunicorn, we need to pattern-match only on the wsgi
process we want to stop or restart.

We are also tuning gunicorn for quads-server to accomodate an 8vcpu VM
system, the defaults are a little low.

Also let's add more instructions in RPM spec for upgrade scenarios.

Fixes: https://github.com/redhat-performance/quads/issues/510
Change-Id: If239488f38b8f06edc0637a85291db99788ad14e